+++ src/react/components/table/private/XUITableBodyRow.tsx
@@ -123,6 +123,7 @@ class XUITableBodyRow<RD extends RowData = RowData> extends React.PureComponent<
               <XUIEditableTableCellReadOnly
                 cellProps={generateCellProps(
                   cell,
+                  column,
                   columnIndex,
                   columns,
                   isCellClickable,
@@ -132,7 +133,6 @@ class XUITableBodyRow<RD extends RowData = RowData> extends React.PureComponent<
                   hasOverflowMenu,
                   this.state.hasCellPrecedence,
                 )}
-                inlineAlignment={column.props.inlineAlignment}
                 key={nanoid()}
                 onPointerOut={() => isCellClickable && this.removeCellPrecedence()}
                 onPointerOver={() => isCellClickable && this.addCellPrecedence()}