+++ src/react/components/table/private/XUITableHead.tsx
@@ -78,17 +78,12 @@ class XUITableHead<RD extends RowData> extends React.PureComponent<Props<RD>> {
               }
 
               const cells = React.Children.toArray(column.props.head) as Array<XUITableCell<RD>>;
-              const isEndAligned =
-                columnIndex === columns.length - 1 && !hasOverflowMenu ? 'end' : 'start';
-              const headingCellAlignment = column.props.inlineAlignment
-                ? column.props.inlineAlignment
-                : isEndAligned;
-
               return cells.map(head => (
                 <XUITableHeadingCell
                   activeSortKey={activeSortKey}
                   generatedCellProps={generateCellProps(
                     head,
+                    column,
                     columnIndex,
                     columns,
                     Boolean(head.props.onCellClick),
@@ -99,7 +94,6 @@ class XUITableHead<RD extends RowData> extends React.PureComponent<Props<RD>> {
                     undefined,
                   )}
                   head={head}
-                  inlineAlignment={headingCellAlignment}
                   isSortAsc={isSortAsc}
                   key={nanoid()}
                   onSortChange={onSortChange}